changed lat and long migration files

+++ db/migrate/20130716131323_add_latitude_to_user.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToUser < ActiveRecord::Migration
   def change
-    add_column :users, :latitude, :double
+    add_column :users, :latitude, :decimal
   end
 end
