+++ db/migrate/20130716131356_add_longitude_to_user.rb
@@ -1,5 +1,5 @@
 class AddLongitudeToUser < ActiveRecord::Migration
   def change
-    add_column :users, :longitude, :double
+    add_column :users, :longitude, :decimal
   end
 end
